support multiple init executions

+++ get-chrome-rpm.sh
@@ -1,4 +1,13 @@
 #!/bin/bash
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm
-yum install -y google-chrome-stable_current_x86_64.rpm
-rm -f google-chrome-stable_current_x86_64.rpm
+. utils.sh
+
+chrome=$(rpm -qa | grep chrome) 
+if [ "$chrome" != ""]; then 
+	log chrome already installed
+else
+	log Installing chrome
+	wget https://dl.google.com/linux/direct/google-chrome-stable_current_x86_64.rpm
+	yum install -y google-chrome-stable_current_x86_64.rpm
+	rm -f google-chrome-stable_current_x86_64.rpm
+	log Finished
+fi
